packages/common/pyproject.toml
[project]
name = "runtime-common"
version = "0.1.0"
description = "Shared library for agents-runtime services and base images"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "httpx>=0.27",
    "structlog>=24.4",
    "pyjwt[crypto]>=2.9",
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28",
    "redis[hiredis]>=7.4.0",
    "argon2-cffi>=25.1.0",
    "uhashring>=2.4",
    "tenacity>=9.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/runtime_common"]
